Split add tracer module
This splits the Add module into LondonAdd and ShanghaiAdd, with the
latter being responsible for the assembly module.
