+++ arithmetization/src/main/java/net/consensys/linea/zktracer/module/hub/Hub.java
@@ -203,7 +203,7 @@ public abstract class Hub implements Module {
   // stateless modules
   @Getter private final Wcp wcp = new Wcp();
 
-  private final Add add = new Add();
+  private final Add add = setAdd();
   private final Bin bin = new Bin();
   private final Blockhash blockhash = new Blockhash(this, wcp);
   @Getter private final Euc euc = new Euc(wcp);
@@ -1058,6 +1058,8 @@ public abstract class Hub implements Module {
   public Address coinbaseAddressOfRelativeBlock(final int relativeBlockNumber) {
     return blockStack.getBlockByRelativeBlockNumber(relativeBlockNumber).coinbaseAddress();
   }
+
+  protected abstract Add setAdd();
 
   protected abstract GasCalculator setGasCalculator();
